fix: cleanup docker images
