--- app/Jobs/InstanceDockerCleanupJob.php
@@ -1,39 +1,0 @@
-<?php
-
-namespace App\Jobs;
-
-use App\Models\Server;
-use Illuminate\Bus\Queueable;
-use Illuminate\Contracts\Queue\ShouldQueue;
-use Illuminate\Foundation\Bus\Dispatchable;
-use Illuminate\Queue\InteractsWithQueue;
-use Illuminate\Queue\SerializesModels;
-use Illuminate\Support\Facades\Log;
-
-class InstanceDockerCleanupJob implements ShouldQueue
-{
-    use Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
-    public $timeout = 500;
-    /**
-     * Create a new job instance.
-     */
-    public function __construct()
-    {
-        //
-    }
-
-    /**
-     * Execute the job.
-     */
-    public function handle(): void
-    {
-        try {
-            $servers = Server::all();
-            foreach ($servers as $server) {
-                instant_remote_process(['docker image prune -f'], $server);
-            }
-        } catch (\Exception $e) {
-            Log::error($e->getMessage());
-        }
-    }
-}
